fix: accommodate compound conditionals when rewriting factories
This patch updates the logic used when rewriting a factory that uses a
ternary to conditionally pull a service, and uses a compound conditional
to do so. Previously, the behavior led to issues whereby any parts of
the conditional prior to testing for the service were dropped, which led
to false positive detections for the legacy service.

These are now rewritten to if/elseif/else statements to ensure behavior
will be the same as previous results.