refactor: Rename "update" to "migrate"
"migrate" better details what the command is doing in this particular
case: "migrating" a piece of code to work directly with Laminas.

Additionally, migration of a third-party library doesn't particularly
care what the name is, but _does_ care about the path being migrated; as
such, I've created an extension to `Repository` named
`ThirdPartyRepository` which inverts the two constructor arguments.

Finally, when testing locally, I found the command would traverse the
`vendor` directory when present, which you generally do not want in this
scenario. As such, I overrode the `files()` method in
`ThirdPartyRepository` to omit anything in the `vendor` directory.