fix: require packages using --dev when pertinent
`composer why -r` includes the `__root__` package, which in turn
indicates if the dependency being queried was required "(for
development)" (i.e., the upper-most package requiring it was for
development).

This patch now runs that command for each package, in order to determine
if the package is for development, or production.

Packages are then required in two batches, one for production, and the
other for development. In either case, if no packages are present, the
operation is skipped and considered successful.

The return status of the command is now the combination of the return
status of both operations, so if one fails or causes errors, the other
will still run, but the command will be in a failure state.