feat: configurable directory for local-test checkouts
I've typically run this from the directory where I build the Satis repo,
and that means I end up getting all the various repositories checked out
in that location, which is not ideal.

This patch updates the command to allow specifying a `--path` or `-p`
argument. When present, that path will be used instead of `getcwd()`.
The value defaults to `getcwd()`.